Convert service color to ES6 module
