+++ app/assets/javascripts/student_profile/record_service.jsx
@@ -1,10 +1,9 @@
 import {merge} from '../helpers/react_helpers.jsx';
+import serviceColor from '../student_profile/service_color.js';
+
 (function() {
   window.shared || (window.shared = {});
-
   const Datepicker = window.shared.Datepicker;
-  const serviceColor = window.shared.serviceColor;
-
   const ProvidedByEducatorDropdown = window.shared.ProvidedByEducatorDropdown;
 
   const styles = {
